Implement some basic debug hooks
Previously, debugging WebAssembly code with the JIT enabled was near
impossible. The vanilla WABT interpreter allows WebAssembly code
execution to be suspended after executing an arbitrary number of
instructions, but it was not possible to reproduce this functionality
for JITted code without incurring a significant performance penalty.

To fill this gap, a number of hook functions have been added which allow
user-provided code to be run when certain events occur. Since these hook
functions are called from the interpreter rather than returning control
back to the caller, it is possible for JITted code to support calling
these hooks. This provides the basic groundwork for building debuggers
and other introspection tools that can run while the JIT is enabled.

Currently, there are hooks for 4 different events: after entering a new
frame on a call, before exiting a frame with a return, before execution
of all instructions, and before trapping. To make use of this
functionality, the user must explicitly opt into it by setting the
appropriate flags in a TraceAbilities struct before constructing the
WebAssembly environment. These flags allow the JIT to avoid emitting
calls into hook functions that will never be used, as these calls do
incur a performance penalty.