Fix offsets in stack trace tests
Previously, the exact PC of the top frame of a stack trace after
trapping was not very well defined. Usually (though not always) it would
point to the instruction *after* the instruction which actually
triggered the trap.

With the addition of debug hooks, these have now been standardized and
should always point to the start of the instruction which caused the
trap. To reflect this, the stack trace test has been updated to contain
the expected offsets.