Add a --test-hooks option to wasm-interp
A new option (--test-hooks) has been added to wasm-interp for the
purposes of testing the new debug hook functionality. Specifying this
option will result in hooks being installed for all supported events
which simply print the type of hook alongside the address of the
instruction that triggered the hook.

This option is meant only to test this functionality and is not meant to
be used outside of the tests, so this option has been deliberately left
out of the help text to discourage its use.